@@ -1,6 +1,6 @@
 {
   "name": "ang-bell-app",
-  "version": "0.6.12",
+  "version": "0.6.15",
   "license": "AGPL-3.0",
   "scripts": {
     "ng": "ng",
@@ -25,18 +25,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.0.3",
-    "@angular/cdk": "^7.0.3",
-    "@angular/common": "7.0.3",
-    "@angular/compiler": "7.0.3",
-    "@angular/core": "7.0.3",
-    "@angular/forms": "7.0.3",
-    "@angular/material": "^7.0.3",
-    "@angular/platform-browser": "7.0.3",
-    "@angular/platform-browser-dynamic": "7.0.3",
-    "@angular/platform-server": "7.0.3",
-    "@angular/router": "7.0.3",
-    "@angular/service-worker": "7.0.3",
+    "@angular/animations": "7.1.3",
+    "@angular/cdk": "^7.1.1",
+    "@angular/common": "7.1.3",
+    "@angular/compiler": "7.1.3",
+    "@angular/core": "7.1.3",
+    "@angular/forms": "7.1.3",
+    "@angular/material": "^7.1.1",
+    "@angular/platform-browser": "7.1.3",
+    "@angular/platform-browser-dynamic": "7.1.3",
+    "@angular/platform-server": "7.1.3",
+    "@angular/router": "7.1.3",
+    "@angular/service-worker": "7.1.3",
     "@covalent/markdown": "^1.0.1",
     "@covalent/text-editor": "^1.0.0",
     "@types/cropperjs": "^1.1.2",
@@ -56,9 +56,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.6",
-    "@angular/cli": "^7.0.5",
-    "@angular/compiler-cli": "7.0.3",
-    "@angular/language-service": "7.0.3",
+    "@angular/cli": "^7.1.3",
+    "@angular/compiler-cli": "7.1.3",
+    "@angular/language-service": "7.1.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
