@@ -108,7 +108,8 @@ export class ExamsAddComponent implements OnInit {
   }
 
   addExam(examInfo) {
-    this.couchService.post(this.dbName, { createdDate: Date.now(), ...examInfo, updatedDate: Date.now() }).subscribe((res) => {
+    const date = this.couchService.datePlaceholder;
+    this.couchService.updateDocument(this.dbName, { createdDate: date, ...examInfo, updatedDate: date }).subscribe((res) => {
       this.documentInfo = { _id: res.id, _rev: res.rev };
       let routerParams = {};
       if (this.examType === 'courses') {
