@@ -128,19 +128,11 @@ export class FeedbackComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   closeFeedback(feedback: any) {
-    const updateFeedback =  { ...feedback, 'closeTime': Date.now(),  'status': 'Closed' };
-    this.couchService.put(this.dbName + '/' + feedback._id, updateFeedback).subscribe((data) => {
-      this.planetMessageService.showMessage('You closed this feedback.');
-      this.getFeedback();
-    },  (err) => console.log(err));
+    this.feedbackService.closeFeedback(feedback).subscribe(() => this.getFeedback());
   }
 
   openFeedback(feedback: any) {
-    const updateFeedback =  { ...feedback, closeTime: '',  status: 'Reopened' };
-    this.couchService.put(this.dbName + '/' + feedback._id, updateFeedback).subscribe((data) => {
-      this.planetMessageService.showMessage('You re-opened this feedback.');
-      this.getFeedback();
-    },  (err) => console.log(err));
+    this.feedbackService.openFeedback(feedback).subscribe(() => this.getFeedback());
   }
 
   goBack() {
